+++ public/classes/models/listen/Event.js
@@ -14,4 +14,5 @@ export const ListerEventLi = {
   o:         { nokey: 'openActiveLink' },
   g:         { nokey: 'goLink' },
   a:         { nokey: 'splitLink' },
+  t:         { nokey: 'openNaturePanel' },
 }