+++ public/classes/models/listen/Project.js
@@ -1,0 +1,6 @@
+import { ItemLi } from './Item.js'
+
+export const ProjectLi = {
+  ...ItemLi,
+  t: { nokey: 'openNaturePanel' },
+}